fixing compiler warning

+++ arena2d-sim/arena/Arena_update.cpp
@@ -282,7 +282,7 @@ void Arena::update()
 				_csvWriter.writeHeader(names);
 			}
 			if(values.size() != _csvWriter.getNumCols()){
-				ERROR_F("Number of metrics changed to %d (before: %d)!", values.size(), _csvWriter.getNumCols());
+				ERROR_F("Number of metrics changed to %d (before: %d)!", (int)values.size(), _csvWriter.getNumCols());
 			}
 			_csvWriter.write(values);
 			_csvWriter.flush();
